+++ src/components/flyout/flyout.test.tsx
@@ -2,7 +2,10 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 
-import { clearSelected, downloadSelected } from '@/features/selected-slice';
+import {
+  clearSelected,
+  downloadSelected,
+} from '@/features/slices/selected-slice';
 
 import { Flyout } from './flyout';
