+++ src/components/index.tsx
@@ -3,10 +3,10 @@ export { CardList } from './card-list/card-list';
 export { Loader } from './loader/loader';
 export { Search } from './search/search';
 export { Description } from './description/description';
-export { Home } from './home/home';
 export { About } from '../pages/about';
 export { Header } from './header/header';
-export { NotFound } from '../pages/404';
+export { NotFound } from '../pages/not-found';
 export { Selected } from './selected/selected';
 export { Flyout } from './flyout/flyout';
 export { ThemeToggle } from './theme-toggle/theme-toggle';
+export { LeftRight } from './left-right/left-right';
